requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.3
django-admin-thumbnails==0.2.4
django-ckeditor==6.2.0
django-crispy-forms==1.14.0
django-encrypted-id==0.3.2
django-htmx==1.12.1
django-image-optimizer==1.0.1
django-jazzmin==2.4.9
django-js-asset==2.0.0
gunicorn==20.1.0
idna==3.3
Pillow==9.0.1
psycopg2-binary==2.9.3
pycrypto==2.6.1
python-resize-image==1.1.20
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tinify==1.5.2
urllib3==1.26.8
